linux: fix broken tests
Fixes errors for comparisons between signed and unsigned ints.

Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/1642361

+++ src/client/linux/minidump_writer/linux_ptrace_dumper_unittest.cc
@@ -426,7 +426,7 @@ TEST_F(LinuxPtraceDumperChildTest, FileIDsMatch) {
 #undef TestBody
 
 TEST(LinuxPtraceDumperTest, VerifyStackReadWithMultipleThreads) {
-  static const int kNumberOfThreadsInHelperProgram = 5;
+  static const size_t kNumberOfThreadsInHelperProgram = 5;
 
   pid_t child_pid = SetupChildProcess(kNumberOfThreadsInHelperProgram);
   ASSERT_NE(child_pid, -1);
@@ -484,7 +484,7 @@ TEST(LinuxPtraceDumperTest, VerifyStackReadWithMultipleThreads) {
 }
 
 TEST_F(LinuxPtraceDumperTest, SanitizeStackCopy) {
-  static const int kNumberOfThreadsInHelperProgram = 1;
+  static const size_t kNumberOfThreadsInHelperProgram = 1;
 
   pid_t child_pid = SetupChildProcess(kNumberOfThreadsInHelperProgram);
   ASSERT_NE(child_pid, -1);
